Implemented backlog for failed patch note posts.
A list has been implemented within requests for managing the Titles and URLs of patch note entries that have not been successfully sent, such as in situations where connection is briefly lost. At the start of every check, any previously unsuccessful posts will try to be sent again until it is successful.