Fixed issue with hard-coded rate limit checks
When posting a message to Discord using PostMessage, a check was previously introduced after sending a POST request to the specified webhook which checks the values of a `X-RateLimit-Remaining` and `X-RateLimit-Reset-After` header in the response to check if the bot is about to be rate limited and wait for a specific amount of time as needed. A check has now been added to ensure they exist within the response returned before attempting to reference them in response to problems recently encountered running the bot.

Because these headers were previously hard-coded into the method on the basis that they would always be returned (assuming the initial POST request was successful); when in reality it is stated in the Discord Developer Documentation that this should not be the case (https://discord.com/developers/docs/topics/rate-limits), it ended up being observed in practice recently that because the bot eventually attempted to reference the value of a header in the response that did not actually exist that time, that it would cause the error to be caught as if it was down to a failure to connect. This led to the patch notes in question beingadded to the failed posts list despite being successfully posted, and then continuing to be repeatedly posted each time the bot continues to check from that point onwards until it is terminated, as this process would lead to the same outcome when the bot went to reference those same headers.