Added pruning of Champion page links
Because of an update made to the League of Legends website, HTML for a Character Card Grid has seemingly been introduced which is now picked up during scraping. This has led to hundreds of Champion page links being posted at once, breaking intended functionality and flooding Users with notifications if a Role has been set. To remedy this, an extra method has been added to the pruning process to remove entries whose URL suggests it is a Champion page rather than a Patch Note post, preventing this problem from occuring again in the future.